Split app into calculators pages
