+++ src/app/app.component.scss
@@ -1,8 +1,13 @@
-.app-layout {
+.layout {
   height: 100%;
 
-  cp-hud {
-    justify-self: start;
+  > * {
+    position: absolute;
+    height: 100%;
     width: 100%;
   }
+}
+
+.router-container {
+
 }
