+++ src/app/pages/page-calculators/page-calculators.component.html
@@ -1,0 +1,10 @@
+<div class="layout stack">
+  <ng-container [ngSwitch]="calculatorType">
+
+    <cp-page-dv-planner *ngSwitchCase="calculators.DvPlanner"></cp-page-dv-planner>
+
+    <cp-page-signal-check *ngSwitchCase="calculators.SignalCheck"></cp-page-signal-check>
+  </ng-container>
+
+  <cp-hud></cp-hud>
+</div>
