+++ src/app/pages/page-calculators/page-calculators.component.scss
@@ -1,0 +1,12 @@
+:host {
+  color: white;
+}
+
+.layout {
+  height: 100%;
+
+  cp-hud {
+    justify-self: start;
+    width: 100%;
+  }
+}
