+++ src/app/pages/page-calculators/page-calculators.component.spec.ts
@@ -1,0 +1,25 @@
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
+//
+// import { PageCalculatorsComponent } from './page-calculators.component';
+//
+// describe('PageCalculatorsComponent', () => {
+//   let component: PageCalculatorsComponent;
+//   let fixture: ComponentFixture<PageCalculatorsComponent>;
+//
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ PageCalculatorsComponent ]
+//     })
+//     .compileComponents();
+//   });
+//
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(PageCalculatorsComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
+//
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
